+++ get-next-line/get_next_line.h
@@ -15,23 +15,20 @@
 
 # include <stdlib.h>
 # include <unistd.h>
-# include <limits.h> // OPEN_MAX 사용
+# include <limits.h>
 
 # ifndef BUFFER_SIZE
 #  define BUFFER_SIZE 42
 # endif
 
-// get_next_line.c Functions
-int		is_there_nl(char *s);
+ssize_t	is_there_nl(char *s);
 char	*get_ret_line(char *store);
-void	mid_process(char *buffer, char *store, size_t read_size);
+void	mid_process(char *buffer, char *store, ssize_t read_size);
 char	*get_next_line(int fd);
-
-// get_next_line_utils.c Functions
 void	free_str(char *str);
-size_t	ft_strlen(const char *str);
-char	*ft_strndup(char *s, size_t len);
-void	*ft_memcpy(void *dst, void *src, size_t size);
+ssize_t	ft_strlen(const char *str);
+char	*ft_strndup(char *s, ssize_t len);
+void	*ft_memcpy(void *dst, void *src, ssize_t size);
 char	*ft_strjoin(char const *s1, char const *s2);
 
 #endif
